Moved the keys logic to utils, simplified the code and removed unused create_genesis func
